+++ CHANGES.md
@@ -7,6 +7,8 @@
 - Fix login form style for `classic` theme
 - Properly handle OGP metadata that doesn't have an image
 - Fix TypeError which prevents logging out
+- Fix auto zoom in when input message in ios safari
+- Add a service discovery browser to the settings modal
 
 ## 11.0.0 (2025-05-21)