+++ CHANGES.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## 11.0.0 (Unreleased)
+## 11.0.1 (2025-06-09)
 
+- #2938: Add a service discovery browser
 - #3672: Images not rendering
 - #3676: Flyout box is not positioned correctly on mobile devices in "overlayed" mode
 - #3689: Chat view not accesible for certain widths in "fullscreen" mode
@@ -11,8 +12,8 @@
 - Properly handle OGP metadata that doesn't have an image
 - Fix TypeError which prevents logging out
 - Fix auto zoom in when input message in ios safari
-- Add a service discovery browser to the settings modal
 - Add a modal to view blocked XMPP addresses
+- Check for PubSub `config-node-max` feature before using the `max` value
 
 ## 11.0.0 (2025-05-21)
 
@@ -29,6 +30,7 @@
 - #1195: Add actions to quote and copy messages
 - #1303: Display non-contacts who sent us a message somehow in fullscreen 
 - #1349: XEP-0392 Consistent Color Generation
+- #1521: Display locked domain behind login field?
 - #1700: Deleted pending contacts reappear after page reload
 - #1810: Create clickable link to load older MAM messages if there is no scrollbars.
 - #2118: Show reflected message in MUC